SAM-14, Оптимизация
Изменения, связанные с переиспользованием collectionView для filters и tune tools
