+++ PhotoTune/Modules/EditingScreen/Views/Tools/Tune/TuneToolCollectionViewCell.swift
@@ -1,0 +1,14 @@
+//
+//  TuneToolCollectionViewCell.swift
+//  PhotoTune
+//
+//  Created by Mikhail Medvedev on 18.12.2019.
+//  Copyright © 2019 Mikhail Medvedev. All rights reserved.
+//
+
+import UIKit
+
+final class TuneToolCollectionViewCell: UICollectionViewCell
+{
+	static let identifier = "TuneToolCollectionViewCell"
+}
