--- PhotoTune/Modules/EditingScreen/Views/Tools/Tune/TuneView.swift
@@ -1,14 +1,0 @@
-//
-//  TuneView.swift
-//  PhotoTune
-//
-//  Created by Mikhail Medvedev on 18.12.2019.
-//  Copyright © 2019 Mikhail Medvedev. All rights reserved.
-//
-
-import UIKit
-
-final class TuneView: UIView
-{
-//implement
-}
